+++ kazoo/exceptions.py
@@ -47,6 +47,21 @@ class KazooTimeoutError(Exception):
     pass
 
 
+class ForceRetryError(Exception):
+    """Raised when some recipe logic wants to force a retry."""
+
+
+class RetryFailedError(KazooException):
+    """Raised when retrying an operation ultimately failed, after
+    retrying the maximum number of attempts.
+    """
+
+
+class InterruptedError(RetryFailedError):
+    """Raised when the retry is forcibly interrupted by the interrupt
+    function"""
+
+
 def _invalid_error_code():
     raise RuntimeError('Invalid error code')
